pair programming with clement
